+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/database/internal/ChatDatabase.kt
@@ -29,6 +29,7 @@ import io.getstream.chat.android.offline.repository.database.converter.internal.
 import io.getstream.chat.android.offline.repository.database.converter.internal.MapConverter
 import io.getstream.chat.android.offline.repository.database.converter.internal.MemberConverter
 import io.getstream.chat.android.offline.repository.database.converter.internal.ModerationDetailsConverter
+import io.getstream.chat.android.offline.repository.database.converter.internal.PrivacySettingsConverter
 import io.getstream.chat.android.offline.repository.database.converter.internal.QuerySortConverter
 import io.getstream.chat.android.offline.repository.database.converter.internal.ReactionGroupConverter
 import io.getstream.chat.android.offline.repository.database.converter.internal.SetConverter
@@ -68,7 +69,7 @@ import io.getstream.chat.android.offline.repository.domain.user.internal.UserEnt
         CommandInnerEntity::class,
         SyncStateEntity::class,
     ],
-    version = 73,
+    version = 74,
     exportSchema = false,
 )
 @TypeConverters(
@@ -83,6 +84,7 @@ import io.getstream.chat.android.offline.repository.domain.user.internal.UserEnt
     MemberConverter::class,
     ModerationDetailsConverter::class,
     ReactionGroupConverter::class,
+    PrivacySettingsConverter::class,
 )
 internal abstract class ChatDatabase : RoomDatabase() {
     abstract fun queryChannelsDao(): QueryChannelsDao