+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/user/internal/UserEntity.kt
@@ -54,6 +54,7 @@ internal data class UserEntity(
     val updatedAt: Date? = null,
     val lastActive: Date? = null,
     val invisible: Boolean = false,
+    val privacySettings: PrivacySettingsEntity?,
     val banned: Boolean = false,
     val mutes: List<String> = emptyList(),
     val extraData: Map<String, Any> = emptyMap(),